Fix NavBar not displaying contents right and login/signup flicker.
